requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
autopep8==1.5.6
backcall==0.2.0
billiard==3.6.4.0
black==21.5b1
celery==5.0.5
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
decorator==5.0.7
Django==3.2
djangorestframework==3.12.4
idna==2.10
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kombu==5.0.2
mpmath==1.2.1
mypy-extensions==0.4.3
parso==0.8.2
pathspec==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.18
ptyprocess==0.7.0
pycodestyle==2.7.0
Pygments==2.8.1
pytz==2021.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
sympy==1.8
toml==0.10.2
traitlets==5.0.5
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
yapf==0.31.0
